Split composer attachment tiles
