+++ scripts/design-token-baseline.txt
@@ -4,7 +4,7 @@
 # lower numbers are better — re-baseline after migrating a screen.
 app/src/main/java/com/pocketshell/app/composer/MarkdownText.kt 4
 app/src/main/java/com/pocketshell/app/composer/PromptComposerQueueBanners.kt 2
-app/src/main/java/com/pocketshell/app/composer/PromptComposerSheet.kt 23
+app/src/main/java/com/pocketshell/app/composer/PromptComposerSheet.kt 20
 app/src/main/java/com/pocketshell/app/conversation/ConversationMessageTurn.kt 1
 app/src/main/java/com/pocketshell/app/conversation/ConversationTextSection.kt 3
 app/src/main/java/com/pocketshell/app/crash/CrashReportsScreen.kt 2
