+++ scripts/file-size-hygiene-baseline.txt
@@ -5,7 +5,7 @@
 # Ratchet rule: baselined files may shrink or disappear, but may not grow.
 # New non-exempt tracked files above threshold_bytes fail the guard.
 # Exemptions: vendor/generated/worktree dirs and shared/core-terminal.
-142988 app/src/main/java/com/pocketshell/app/composer/PromptComposerSheet.kt
+134030 app/src/main/java/com/pocketshell/app/composer/PromptComposerSheet.kt
 207337 app/src/main/java/com/pocketshell/app/composer/PromptComposerViewModel.kt
 139372 app/src/main/java/com/pocketshell/app/projects/FolderListScreen.kt
 180197 app/src/main/java/com/pocketshell/app/projects/FolderListViewModel.kt
